porttester/requirements.txt
streamlit==1.20.0
pyyaml==6.0
pandas